Add a javascript runner for running javascript before and after requests
This implement a basic javascript runner.
* support Response API
* support set/get Request API variables
* support set/get client global variables
* support jsonPath query in javascript

Dev notes:
A intermediate file is created for each request in the script folder,
`last_javascript.mjs` is the last javascript file that was run.
This is useful for debugging. For instance you can run the file from the
command line `node --inspect-brk last_javascript.mjs` and then open up
`chrome://inspect` in chrome based browers and debug the javascript.